+++ src/main/java/codesquad/domain/user/UserRegistrationHandler.java
@@ -5,7 +5,7 @@ import codesquad.server.handler.annotation.Handler;
 import codesquad.server.handler.annotation.HttpMethod;
 import codesquad.server.http.HttpRequest;
 import codesquad.server.http.HttpResponse;
-import static codesquad.server.util.File.readFileContent;
+import static codesquad.server.util.FileUtils.readFileContent;
 
 @Handler("/registration")
 public class UserRegistrationHandler extends CustomRequestHandler {
