+++ src/main/java/codesquad/server/handler/AbstractRequestHandler.java
@@ -2,11 +2,9 @@ package codesquad.server.handler;
 
 import codesquad.server.http.ContentType;
 import codesquad.server.http.HttpResponse;
-import codesquad.server.util.File;
-import java.io.InputStream;
+
 import static codesquad.server.handler.Constants.ERROR_PAGE_PATH;
-import static codesquad.server.util.File.readFileContent;
-import static codesquad.server.util.File.readInputStreamToByteArray;
+import static codesquad.server.util.FileUtils.readFileContent;
 
 public abstract class AbstractRequestHandler implements RequestHandler {
     protected HttpResponse.Builder ok(byte[] body) {
