+++ src/main/java/codesquad/server/handler/StaticFileHandler.java
@@ -4,7 +4,7 @@ import codesquad.server.http.ContentType;
 import codesquad.server.http.HttpRequest;
 import codesquad.server.http.HttpResponse;
 import java.util.Map;
-import static codesquad.server.util.File.loadStaticFiles;
+import static codesquad.server.util.FileUtils.loadStaticFiles;
 
 public class StaticFileHandler extends AbstractRequestHandler {
     private final Map<String, byte[]> staticFiles;
